Relocate googletests for compile time functionality into their own file
